sublime-config: Change GOPATH/GOROOT checker to raise exceptions
When checking GOROOT and GOPATH to ensure they exist on disk, raise
an exception if they are not found. This allows consuming packages
to present the user with more information about why the environment
variable value could not be found.

This change also syncs the changelog and in-code version numbers
with the tags that have been used so far.

Change-Id: I00ee05bde751848737d0346193444cece5074089
Reviewed-on: https://go-review.googlesource.com/16570
Reviewed-by: Jason Buberel <jbuberel@google.com>